Update DB setup script
I've set this up so that we can run this before starting the app for the
first time to get some needed data in the database. Feel free to add
tables, columns and fields to this so that we can all be working with a
consistent database.

In a previous commit I set hibernate to only update/alter tables if they
do not match the definition in the classes. Before it was just nuking
the the whole DB. This way, we won't lose all user data on every restart
of tomcat.